Move pods into common directory.
